Tests are moved into another package
